requirements.txt
Django==1.5.1
South==0.7.6
django-debug-toolbar==0.9.4
django-jsonfield==0.9.3
django-registration==0.9
djangorestframework==2.2.5
